# [LTS 9.2] perf: Disallow mis-matched inherited group reads