@thefossguy-ciq 2025-07-30
> Do we want to have the kABI fix patch as a separate commit or should just fix it in the commit we cherry-picked? For now I added as a separate one for everyone to review, will merge it if required

I don't know if we have an official consensus about this but so far, I've seen it done in a single commit and specify the changes using `upstream-diff`.