@bmastbergen 2025-07-30
> ## 🔍 Upstream Linux Kernel Commit Check
> * ⚠ PR commit `8b7eacd9347a (perf: Disallow mis-matched inherited group reads)` references upstream commit
>   `32671e3799ca` which has been referenced by a `Fixes:` tag in the upstream
>   Linux kernel:
> 
> ```
>     a71ef31485bb perf/core: Fix potential NULL deref (Peter Zijlstra)
> ```
> 
> _This is an automated message from the kernel commit checker workflow._

This wasn't automated, I ran my upstream commit checker script locally