@bmastbergen 2025-07-31
> Do we want to have the kABI fix patch as a separate commit or should just fix it in the commit we cherry-picked? For now I added as a separate one for everyone to review, will merge it if required

Sorry I didn't comment on this earlier.  As Pratham said, usually we just do one commit with an upstream-diff explaining the macro usage.  I think that would be my preference here.