@shreeya-patel98 2025-07-31
> > _This is an automated message from the kernel commit checker workflow._
> 
> I think we should pick this up as well. Usually we'd tag this with 'cve-bf [CVE-2023-5717](https://github.com/advisories/GHSA-gv6g-gf42-rjg7)'

So your script checks if there were any fixes upstreamed for this particular patch. That's cool :)